Organize Folder Structure + split client things
