--- src/components/editor/input_manager/event-handling.tsx
@@ -1,34 +1,0 @@
-import { BaseNode } from "~/components/nodes/base-node";
-import { NodeSlot } from "~/components/nodes/slot/node-slot";
-
-export enum InputEvents {
-    POINTER_MOVING,
-    CLICK_ON_NODE,
-    CLICK_ON_NODE_SLOT,
-    HOVER_NODE,
-    HOVER_SLOT,
-    HOVER_BACKGROUND
-}
-
-export interface EventData {
-    event?: UIEvent
-    node?: BaseNode,
-    slot?: NodeSlot
-}
-
-export interface EventHandlerInterface {
-    handler_func: (event_data: EventData) => void;
-    name: string
-}
-
-export class EventHandler {
-    name: string;
-    private func: (event_data: EventData) => void;
-
-    constructor(handler_name: string, handler_func: (event_data: EventData) => void) {
-        this.name = handler_name;
-        this.func = handler_func;
-    }
-
-    get handler() { return this.func; }
-}
