+++ src/editor/tools/selection-tool.tsx
@@ -1,0 +1,97 @@
+import { BaseEditorTool } from "./base-tool";
+import { BaseNode } from "~/wrapper/nodes/base-node";
+import { Vector2 } from "~/wrapper/data_types/geometry";
+import { NodeEditor } from "../node-editor";
+import { NodeSlot } from "~/wrapper/nodes/slot/node-slot";
+import { ConnectionController } from "../../wrapper/controllers/connection-controller";
+import { NodeController } from "~/wrapper/controllers/node-controller";
+import { SelectionController } from "~/editor/controllers/selection-controller";
+import { NodePreview } from "../ui/misc/node-preview";
+
+export class SelectionTool extends BaseEditorTool {
+    selection_controller: SelectionController
+    connection_controller: ConnectionController;
+    node_controller: NodeController;
+    node_editor: NodeEditor
+
+    _selected_preview: NodePreview | null = null;
+
+    constructor(node_editor: NodeEditor) {
+        super(node_editor);
+        this.selection_controller = node_editor.selection_controller;
+        this.connection_controller = node_editor.scene_controller.connection_controller;
+        this.node_controller = node_editor.scene_controller.node_controller;
+        
+        this.node_editor = node_editor;
+    }
+
+    onClickOnNodeSlot(slot: NodeSlot): void {
+        this.connection_controller.select_slot(slot);
+        this.selection_controller.clearSelection();
+    }
+
+    onPointerDown(e: PointerEvent): void {
+        const [screen_pos, world_pos] = this.selection_controller.editor_space.get_cursor_pos(e)
+        if (e.button == 0) {
+            if (this.selection_controller.has_selected) {
+                this.selection_controller.clearSelection();
+            }
+            this.connection_controller.unselect_slot();
+
+            (e.currentTarget as HTMLElement).setPointerCapture(e.pointerId);
+            this.selection_controller.onStartAreaSelection({x: world_pos.x, y: world_pos.y});
+        }
+    }
+
+    onPointerUp(e: PointerEvent): void {
+        this.selection_controller.moving = false;
+        if (this.selection_controller.selecting) {
+            this.selection_controller.stopSelection();
+        }
+    }
+
+    onClickOnNode(node: BaseNode): void {
+        this.connection_controller.unselect_slot();
+        this.selection_controller.onClickOnNode(node)
+    }
+
+    onMoveCursor(pos: Vector2, delta: Vector2, all_nodes: BaseNode[]): void {
+        this.selection_controller.onMoveCursor(
+            pos, 
+            delta,
+            all_nodes
+        )
+
+        if (this.selection_controller.moving) {
+            this.selection_controller.selected_nodes.forEach(node => {
+                node.all_slots.forEach(slot => {
+                    slot.update_anchor();
+                });
+            });
+        }
+    }
+
+    onClickOnNodePreview(node_preview: NodePreview): void {
+        this.selection_controller.selected_node_type = node_preview.node_constructor.type_name;
+        
+        if (this._selected_preview != null) {
+            this._selected_preview.selected = false
+        }
+        node_preview.selected = true;
+        this._selected_preview = node_preview
+    }
+
+    onHoverSlot(slot: NodeSlot): void {
+        this.connection_controller.hovered_slot = slot;
+        this.selection_controller.hovered_node = slot.parent_node;
+    }
+
+    onHoverNode(node: BaseNode): void {
+        this.selection_controller.hovered_node = node;
+    }
+
+    onHoverBackground(): void {
+        this.selection_controller.hovered_node = null;
+        this.connection_controller.hovered_slot = null;
+    }
+}
