+++ src/editor/ui/misc/grid.tsx
@@ -1,0 +1,36 @@
+import { createSignal } from "solid-js";
+import { EditorCamera } from "~/editor/internal/editor-space";
+import { Vector2 } from "~/wrapper/data_types/geometry";
+
+export class Grid {
+    _grid_size: () => Vector2;
+    _set_grid_size: (grid_size: Vector2) => void;
+
+    constructor(grid_size: Vector2) {
+        const [gridSize, setGridSize] = createSignal(grid_size)
+
+        this._grid_size = gridSize;
+        this._set_grid_size = setGridSize
+    }
+
+    get grid() { return this._grid_size() }
+    set grid(grid_size: Vector2) { this._set_grid_size(grid_size) }
+
+    public View(camera: EditorCamera) {
+        return (
+            <div class="grid"
+                style={{
+                    position: "absolute",
+                    margin: 0,
+                    height: "100%",
+                    width: "100%",
+                    "background-position": `${-camera.offset.x * camera.zoom}px ${-camera.offset.y * camera.zoom}px `,
+                    "background-image": "repeating-linear-gradient(#ccc 0 1px, transparent 1px 100%), repeating-linear-gradient(90deg, #ccc 0 1px, transparent 1px 100%)",
+                    "background-size": `${this._grid_size().x * camera.zoom}px ${this._grid_size().y * camera.zoom}px`
+                }}
+            >
+                
+            </div>
+        )
+    }
+}
